Changed the type of srcPort/dstPort from size_t to uint16_t in ConnectionData. It saves 4 bytes per connection.

+++ Packet++/header/TcpReassembly.h
@@ -82,9 +82,9 @@ struct ConnectionData
 	/** Destination IP address */
 	IPAddress* dstIP;
 	/** Source TCP/UDP port */
-	size_t srcPort;
+	uint16_t srcPort;
 	/** Destination TCP/UDP port */
-	size_t dstPort;
+	uint16_t dstPort;
 	/** A 4-byte hash key representing the connection */
 	uint32_t flowKey;
 	/** Start TimeStamp of the connection */
